pipe_plugins: Create new header with data types to be shared between collectors
